WarpX class: move CheckKnownIssues  class to anonymous namespace in WarpXInitData.cpp (#5674)
`CheckKnownIssues` is only used once in `WarpXInitData.cpp` and it is
implemented in this source file. This PR turns `CheckKnownIssues` from a
member function of the WarpX class to a pure function inside an
anonymous namespace in `WarpXInitData.cpp`. In the process, the function
is renamed `CheckKnownEMSolverIssues`, since this is more descriptive of
what the function does.
Elimating `CheckKnownIssues` from `WarpX.H` simplifies the interface of
the WarpX class.

---------

Co-authored-by: Edoardo Zoni <ezoni@lbl.gov>