Remove buggy synchronization at end of steps (#5820)
This PR does cleanup, removing a potential bug that happens when the
velocity is synchronized with position in the final diagnostic. The
particle boundary conditions had not been applied so the field gather
could be out of bounds.

Further cleanup is done, renaming Synchronize to
SynchronizeVelocityWithPosition, and having it check is_synchronized,
since the synchronization may have already been done.

Note that the factor `1e-3` in the code change
```cpp
bool const end_of_step_loop = (step == numsteps_max - 1) || (cur_time + dt[0] >= stop_time - 1.e-3*dt[0]);
```
is a somewhat arbitrary factor to allow for round off error in the time.
If the `cur_time + dt[0]` is almost but not quite equal to `stop_time`,
then that's close enough. Comment copied here from
https://github.com/BLAST-WarpX/warpx/pull/5820#discussion_r2076467690
for better visibility.